requirements.txt
openai>=1.30.0
pyyaml>=6.0
Pillow>=10.0.0
